+++ CHANGELOG.md
@@ -4,6 +4,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+### Added
+- Added support for global template variables
+
 ### Fixed
 - Fixed broken installer
 - Fixed debug mode reporting incorrect callers for queries